backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "mcp>=1.25.0",
    "pageindex>=0.2.4",
    "passlib>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
